+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 * 0.3.2 (2023-11-08): Fix error when a target file does not exist; update
   dependencies
 * 0.3.3 (2023-11-08): Ignore commented commands
+* 0.4.0 (2023-11-10): Add `-v` option and don't print up to date targets; move
+  bunt calls to functions; improve comments and miscellaneous improvements;
+  don't process dependencies for a file target unless needed (forced via `-B`,
+  doesn't exist, or outdated); change default outdated response to false to
+  avoid processing a file target unnecessarily
 
 [#1]: https://github.com/qtfkwk/mkrs/issues/1
 [default `Makefile.md` for a Rust project]: styles/Makefile.rust.md